+++ frag/network/scripts/process_enamine_compounds.py
@@ -47,11 +47,11 @@ import logging
 import os
 import sys
 
-from process_utils import error
-from process_utils import write_supplier_nodes
-from process_utils import write_isomol_nodes
-from process_utils import write_isomol_suppliermol_relationships
-from process_utils import write_nodes
+from frag.utils.process_utils import error
+from frag.utils.process_utils import write_supplier_nodes
+from frag.utils.process_utils import write_isomol_nodes
+from frag.utils.process_utils import write_isomol_suppliermol_relationships
+from frag.utils.process_utils import write_nodes
 
 
 # Configure basic logging
